Convert to Spring Boot and Three Layer Architecture.
First iteration of Three Layer Architecture. TramStopService has also been refactored.
