+++ .gitignore
@@ -1,33 +1,33 @@
-# Project exclude paths
-# git config --global core.excludesfile ~/.gitignore_global
+HELP.md
+target/
+!.mvn/wrapper/maven-wrapper.jar
+!**/src/main/**/target/
+!**/src/test/**/target/
 
-# Some common .gitignore configurations: https://gist.github.com/octocat/9257657
-
-# Copy the next content on the .gitignore_global file:
-
-.DS_Store
-
-*.class
-
-bin/
-.settings/
-
-# Eclipse files
-.project
+### STS ###
+.apt_generated
 .classpath
-.metadata/
+.factorypath
+.project
+.settings
+.springBeans
+.sts4-cache
 
-# IntelliJ
-.idea/
+### IntelliJ IDEA ###
+.idea
+*.iws
 *.iml
 *.ipr
-*.iws
-atlassian-ide-plugin.xml
 
-# Maven files
-target/
+### NetBeans ###
+/nbproject/private/
+/nbbuild/
+/dist/
+/nbdist/
+/.nb-gradle/
+build/
+!**/src/main/**/build/
+!**/src/test/**/build/
 
-# Package Files #
-*.jar
-*.war
-*.ear
+### VS Code ###
+.vscode/
