+++ src/main/java/com/tfgm/TfgmApplication.java
@@ -5,20 +5,22 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.concurrent.TimeUnit;
 
+import com.tfgm.services.TramStopService;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-public final class App {
-  private App() {}
+@SpringBootApplication
+public class TfgmApplication {
 
   public static void main(String[] args)
       throws URISyntaxException, IOException, InterruptedException {
-
+    //    SpringApplication.run(TfgmApplication.class, args);
     int count = 0;
-    TramController tramController =
-        new TramController("src/main/resources/com/tfgm/TramStopData.json");
+    TramStopService tramService =
+        new TramStopService("src/main/resources/static/TramStopData.json");
 
     while (true) {
       System.out.println("---------LOOOOP---------" + count);
-      tramController.update();
+      tramService.update();
       TimeUnit.SECONDS.sleep(10);
       count += 10;
     }
