--- src/test/java/com/tfgm/AppTest.java
@@ -1,14 +1,0 @@
-package com.tfgm;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.junit.jupiter.api.Test;
-
-/** Unit test for simple App. */
-class AppTest {
-  /** Rigorous Test. */
-  @Test
-  void testApp() {
-    assertEquals(1, 1);
-  }
-}
