+++ src/test/java/com/tfgm/TfgmApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tfgm;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TfgmApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
